Remake of Profil_Windows with a better interface and add of an icon app
I did remake the profile windows again because the last one as not
really in good shape. It's still not perfect but it's certainly better
than before. Abandonned the use of a GridBagLayout for a GridLayout
(easier to use and more beautiful). I designed a simple icon for the app
even if i didn't implemented it yet.